[RORDEV-2170] Rename to WarmDockerImageCache; match the substitutor the build uses (review)
